requirements.txt
django-picklefield